move types to own file, move tests to own folder
