+++ packages/components/src/Menu/index.ts
@@ -1,6 +1,3 @@
-export {
-  Menu,
-  type SectionProps,
-  type ActionProps,
-  type MenuProps,
-} from "./Menu";
+export { Menu } from "./Menu";
+
+export type { SectionProps, ActionProps, MenuProps } from "./Menu.types";
